chore: clear the two CodeFactor findings attributable to recent work
test_openms_consensus.py had grown to 1198 lines against a 1000-line limit, all
of it from tests added over the last few PRs. The option/metadata classes
(unassigned PSMs, missed cleavages, enzyme resolution) move to
test_openms_consensus_options.py: they cover the converter's *options* rather
than the core consensusXML -> QPX extraction, so the split follows a real seam
rather than an arbitrary line count. 1198 -> 993.

psm_records_for_pid crossed the complexity limits (52/50 statements, 13/12
branches) when missed_cleavages was added. The additional_scores assembly moves
to _psm_additional_scores; behaviour is unchanged and it returns the site scores
the caller already needed for to_modifications.

Both were consequences of my own changes, and CodeFactor only runs on the
dev -> main diff so neither surfaced on the individual PRs.
